old_code/requirements.txt
streamlit==1.28.1
openai==1.3.7
anthropic==0.7.7
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
